fix(cli): reconcile completion coverage with current main
